Switch `amount` type from `Long` to `Double` in `Invoice` and `Reimbursement` models
- Update models (`Invoice`, `Reimbursement`, `InvoiceCreate`) to use `Double` for `amount` fields.
- Adjust related logic in `InvoiceServiceLive` for calculations and precision handling.
- Refactor reimbursement map and test cases to support `Double` values.
- Expand tests with scenarios involving non-integer reimbursement amounts.

+++ domain/src/scala/domain/models/invoice/Reimbursement.scala
@@ -6,8 +6,8 @@ import zio.json.*
 
 case class Reimbursement(
   driverName: DriverName,
-  totalAmount: Long,
-  to: Map[DriverName, Long]
+  totalAmount: Double,
+  to: Map[DriverName, Double]
 )
 
 object Reimbursement {
@@ -16,5 +16,5 @@ object Reimbursement {
 
   given Schema[Reimbursement] = Schema.derived[Reimbursement]
 
-  given Schema[Map[DriverName, Long]] = Schema.schemaForMap[DriverName, Long](DriverName.encode)
+  given Schema[Map[DriverName, Double]] = Schema.schemaForMap[DriverName, Double](DriverName.encode)
 }
